random quote generator in python
In this example, the quotes variable contains a list of predefined quotes. The get_random_quote function uses the random.choice() function to randomly select a quote from the list and return it.

You can add or remove quotes from the list and also you can create a function that helps you to add more quotes to the list.

This is just a basic example and you can expand it to include more features like the ability to choose quotes by author, topic, or keyword.

You can also store the quotes in a database or a file and read them from there instead of hard-coding them into the program.